+++ src/main/java/gregtech/api/util/item/GhostCircuitItemStackHandler.java
@@ -13,8 +13,8 @@ import gregtech.api.util.GTUtility;
 import gregtech.common.items.ItemIntegratedCircuit;
 
 /**
- * Inventory handler dedicated for {@link gregtech.common.gui.modularui2.sync.GhostCircuitSyncHandler
- * GhostCircuitSyncHandler} and {@link gregtech.common.gui.modularui2.widget.GhostCircuitSlotWidget
+ * Inventory handler dedicated for {@link gregtech.common.modularui2.sync.GhostCircuitSyncHandler
+ * GhostCircuitSyncHandler} and {@link gregtech.common.modularui2.widget.GhostCircuitSlotWidget
  * GhostCircuitSlotWidget}.
  */
 public class GhostCircuitItemStackHandler implements IItemHandlerModifiable {
